requirements.txt
nltk==3.2.5
numpy==1.23.0
Cython==0.28.2
scikit-learn==1.1.1
scipy==1.0.0
torchtext==0.2.3
transformers==4.20.1